mpl_toolkits.Basemap
Basemap does not do any plotting on it’s own, but provides the
facilities to transform coordinates to one of 25 different map
projections (using the PROJ.4 C library). Matplotlib is then used to
plot contours, images, vectors, lines or points in the transformed
coordinates. Shoreline, river and political boundary datasets (from
Generic Mapping Tools) are provided, along with methods for plotting
them. The GEOS library is used internally to clip the coastline and
polticial boundary features to the desired map projection region.